requirements.txt
gunicorn>=20.1.0
uvicorn>=0.15.0
fastapi>=0.70.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
jieba>=0.42.1
pytest>=6.0.0
pytest-html>=3.1.1
python-multipart>=0.0.6
Pillow>=9.0.0
torchvision>=0.11.1
torch>=1.10.0

# RAG pipeline (Mistral API + Tavily + LangGraph)
langchain>=0.3.0
langchain-mistralai>=0.2.0
langgraph>=0.2.0
tavily-python>=0.5.0
